requirements.txt
PixivPy~=3.5.11
cqhttp==1.3.1
imageio==2.6.1
peewee==3.13.1
robobrowser==0.5.3
pygifsicle==1.0.1
selenium==3.141.0
Flask==1.1.1
Flask-Apscheduler==1.12.2

nonebot[scheduler]~=1.6.0
aiocqhttp>=1.2
lxml>=4.4.1
pytz>=2019.3
requests>=2.22.0
sogou_tr_free>=0.0.6
zhconv>=1.4.0
Pillow>=6.2.1
TwitterAPI>=2.5.10
matplotlib>=3.2.0
numpy>=1.18.0
beautifulsoup4>=4.9.0
pygtrie>=2.0
tinydb>=4.0
aiohttp>=3.6
peony-twitter[all]~=1.1.7